Added support for "if you've Warcried in the past 8 seconds"
Added support for this modifier that is found on Berserker's War Bringer ascendancy node. As having warcried recently (the last 4 seconds) also implies having warcried in the past 8 seconds and "conditionUserWarcryRecently" already has the implyCond "UsedSkillRecently", I've also implemented support for multiple implied conditions in the form of a parameter implyCondList.