Implement Pure Agony notable
In practice, the only mod that has an effect on calculations is 'Minions deal x% increased damage while you are affected by a herald. The calculations are properly done for this mod. The maximum amount of virulence stacks required to bump the cap from 40 to 80. We still have no way to selectively increase the cap by whatever mod we have. So people can now set themselves at 60 stacks even without jewels. As for the maximum number or sentinels of purity, this is not listed anywhere, making this mod a noop for us. There is an existing helmet enchant that does not do anything either.