Re-re-update
Rebased once again since it's been a while. Messed up the rebase slightly and accidentally duplicated the DoT stuff in CalcOffence. This fixes it.

Also had to fit DoT changes since my original PR. Had to add an if statement that doesn't overwrite the existing output for the DoT type if the output already exists.