Support Debilitated status effect
Debilitated applies 10% less damage and 20% less movement speed to the enemy.

There are 3 sources of debilitate:
* Natural Authority (notable)
* Punishment
* Phantasmal Fireball

These sources have different chances to apply the effect and for different durations, so just make it a config entry.
Implement it as a 10% less damage taken. It's technically the same as an enemy 10% less damage, except that the latter is not supported in calculations.