Fix Legion Jewel EditedNode Data Saving/Loading
This fixes loading and saving multiple legion jewel conquered passives. Currently only a single modified node saves/loads properly and anything beyond the first node is discarded. Edited nodes that are not conquered or allocated anywhere in the build are pruned when the build is saved, to avoid reintroducing issue #3156 (that was technically fixed by #3229, but with.. various downsides.)