Fix 'defend with % armour' master to not effect total armour value
This mastery was not updated to use the use the `ArmourDefence` and `MAX` node parsing. So it was modifying the flat armour value on a character which is not how it works in game.
Also added the condition check for `LowEnergyShield` and `FullLife` to only appear when needed

Build Code:
https://pastebin.com/QF2sbh7y