Add zoom support by using shift + mouse
This adds support to zoom in and out in any `EditControl`.

Main usage for this is the `Notes` section, which can be kinda hard to read on high resolutions.

Many build authors meanwhile use this area to describe detailed step-by-step leveling guides with a build and having the screen open while reading is probably an regular usecase. Beeing able to zoom in here might help alot of ppl.

For now I only enabled zooming for the notes. Not sure whether other editors (like editing details in the crafting screen would benefit from it also).