Fixed a variety of incorrect catalyst scaling
The changes are based off some testing and work I did in 17/4/2021
So any recently changed mods or new uniques will not be guaranteed to be correct.
https://github.com/PathOfBuildingCommunity/PathOfBuilding/issues/2521#issuecomment-821748654
https://docs.google.com/spreadsheets/d/11mOp3VLPXgbhJq562DdFmP34nx4tNqoF47LbrM-R93g/edit#gid=0

I have verified all of these either using trade or through in game but I haven't collated documentation due to the large number of changes.