FIX: limiting scope to self damage only.
CWDT loops while a really cool feature are hard to support correctly in
POB. Due the amount of assumptions the involved code has to make
regarding the state of the loop, the numbers aren't always accurate.
Staring with this commit the scope is being limited to only add support for
self hit damage as i think it'd be a waste to throw all that code away
and that part should work perfectly fine.