WIP: mixed group setups and two part skills fix
Currently there's an issue with two part skills (active and support)
such as Arcanist Brand where the support part will not be applied to
skills socketed into the same item but in different groups. An example
would be to two groups both socketed into Weapon 1. One having the
Arcane Brand skill gem and the other any other skill. The other skill
will not be supported by Arcane Surge even though they are socketed into
the same item and should be considered as linked.