Psuedo Fix for "ElementalHitSkeletonKnight"
A Scuffed but functional Fix for Vaal Slayer & Newly Added Colossal Vaal Fallen

Method to the fix is to take the inherit "physical" (its not supposed to be physical) damage and then convert 100% of that damage into each element while flagging DealNoPhysical and DealNoChaos to emulate hitting with the skill repeatedly.

there are some problems with this however, such as how this skill is used by both Projectile Spectre aswell as Melee Area Spectre as well as flat phys scaling your damage (it wont in game)
so your damage can get inflated, however this is the best we can do atm until someone more capable comes along to fix this skill properly, untill then this will relatively properly convey your overall dps with Elemental Hit.