Fix Divergent Arcane Cloak
Divergent Arcane Cloak was scaling the added damage by a per stat instead of a percentage stat.
This adds support for using a multiplier variable in place of the percentage value.
Also did some testing in game and Arcane Cloak and other similar effects appear to round up the final value so changed that logic too