New save format
Uses the Path of Exile BaseItemType.Id for the gem in gemId (as it always has been).
The skillId attribute is also kept and maps to the gems variant primary GrantedEffect.Id.
Adds a new attribute called variantId that maps to the variant of the game, eg. Arc, ArcAltX, or ArcAltY. This is to help identify which variant is used rather than relying on every gem variant having its own unique granted effect.

For the majority of gems skillId and variantId are the same, but it is not true for every case.