Sort the CalcBreakdownControl taking in consideration boolean values.
Some breakdowns can contain a mix of `number` and `boolean`. e.g. the skill buffs breakdown if you have a golem active will have a `Condition: have golem` with value of `true`.

The values in this kind of breakdown are pretty sparse and sorting them in a coherent way is not easy, so for the sake of this PR we can just keep skip(return false) when comparing number and boolean.