Basic Charges stats in Calc tab
This commit adds some simple stats for the 3 base types of charges in the Calc tab: Max, Current, Spendable (max-min) and duration.

It would be possible to add stats for all other possible charges, but it would take quite a bit of space, and it would not offer much useful information as apart from the `Charge Duration Mastery` and the `Swift Killer` ascendancy node, there isn't much affecting them.