Fix Law of the Wilds using the wrong minion stats
Law of the Wilds was using the wrong base minion for its stats so the damage was very far off.

It was also using the wrong actor level for the minion and I believe it always grabs the level from the `levelRequirement` stat and the `base_display_minion_actor_level` is only used for the tooltip in game but not the actual minion itself.
The only other minion with differing `base_display_minion_actor_level ` vs `levelrequirement` is Vaal Summon Skeletons. I'll will check this assumption with GGG