FIX: use mirage archer style calcs
Sacred wisps causes the Summon Sacred Wisps skill to be triggered when a
projectile spell is fired from a wand. This triggers 2 wisps which then
have an x percent to use the skill that was used to trigger them when
you use it. They likely do not care about whether you actually hit
something with the projectile or not. They're also likely not a trigger
which means they're not under the same tick rounding contrstains like
other triggers.