Implement wisps support (#7532)
* FEAT: impl wisps support

* FIX: add display stat for summon wisps

* FIX: less damage mod not applying

* FIX: use mirage archer style calcs

Sacred wisps causes the Summon Sacred Wisps skill to be triggered when a
projectile spell is fired from a wand. This triggers 2 wisps which then
have an x percent to use the skill that was used to trigger them when
you use it. They likely do not care about whether you actually hit
something with the projectile or not. They're also likely not a trigger
which means they're not under the same tick rounding contrstains like
other triggers.

* TEST: add a test for Sacred Wisps

* TEST: fix test for sacred wisps

Sacred wisps support is two part so the SkillTriggerRate check will fail

* FIX: add stat maps to .txt skill files

* Separate stat display in sidebar

* Change flag to mirageArcher

---------

Co-authored-by: LocalIdentity <localidentity2@gmail.com>