Fix Frost Blades of Katabasis DoT not scaling with some mods
The DoT portion of the skill was not scaling with the `(70-100)% increased Elemental Damage with Melee Weapons` mod on Tinctures.
It was also not scaling with the repeat damage on Multistrike. I tested it in game and the repeat damage definitely applies to the DoT component.
The PR removes the skill hit tag from the DoT component so it doesn't incorrectly show hit damage and stats
It also removes crit chance from showing up unless a skill part has a hit component