FIX: Blasphemy auras triggering mana cost warning
The warning calculations did not consdier the case of a skill having
both reservation and mana costs. Mana cost would be checked after applying
reservation and not before causing the warning to show up in some cases.

This pr adds a check and handling for such cases, as well as using the
warnings from output to determine color override for cost labels instead
of re-checking pool availability.