FIX: add black background to detail text in character dropdown
Character names are currently drawn over the top of the details on the
right of the dropdown. If a name happens to be very long (often happens
when using tofu for utf-8) the details stop being legible. This adds a
black background behind the detail text to make it more readable.