FIX!: Sacrificial Zeal no longer default on
Initially when trying to fix issue
https://github.com/PathOfBuildingCommunity/PathOfBuilding/issues/8421 i
made the config default enabled to maintain current behaviour of it
being on the moment Hand of the Fervent was equipped. I was under the
impression that the ifFlag would prevent the config application function
from running but that doesn't not seem to be the case. I've tried a few
different ways of fixing this but i've decided that the best course of
action would be to simply make the effect opt in rather than opt out.