FIX: disabling spells preventing curse application (#8522)
The 'Your Spells are disabled' mod found on the Gruthkul's Pelt unique
chest armor currently prevents curse on hit mods from applying. This is
due to the 'chance to Curse Enemies with X on Hit' being implemented in
POB as a simple granted skill without making the distinction between
generic granted skills and on hit (application) skills (curses). This
solution isn't the prettiest but i'd argue it's the most sane given what
information is available to the ModStore.