FIX: incorrect parsing of non-exceptional keyword
Parsing of the Gem-lined Cap Ascendancy Notable mod "+3 to Level on all non-Exceptional Support Gems" was incorrect due to faulty handling of the hyphen causing "non-exceptional" to be split into two keywords: "non" and "exceptional". This caused the mod to not apply as there does not exist a gem tag called "non". This commit fixes this by improving handling of non space characters in the keyword string and implementing handling for the negation part in "non-exceptional" inside of calcLib.gemIsType function.