FIX: first explicit missing if it does not produce a modList (#8576)
The item parsing logic in WIKI mode seems to rely on finding a mod that
produces a modList to set the foundExplicit variable to true which then
allows mods that do not produce a modList to be added to lineList. This
causes issues with items such as Dialla's Malefaction where the first
explicit line does not have any parsing which causes it to not be added
to linesList. Interestingly enough other mods wihch are parsed
piece-wise work fine here as they produce and empty modList with the
extras paramter.