Add support for "per two fortification on you"
NOTE: This doesn't correctly work when putting the item on an AG because
actor = "player" is necessary. actor = "enemy" did also not work and is
less safe. Not specifying an actor would make it based on monster
fortification stacks in some cases. "player" is the best solution until
we have actor = "self"