FIX(mirage): fix cap GC cooldown by mirage+attack delay
The Howlcrack unique item sets the cooldown of warcies to 0 causing the
GC dps to explode into infinity due to division by zero. This was caused
by the assumption that the mirages instantly spawn and attack. This
commit makes it so that the cooldown is at least the amount of time it
takes to spawm them all and attack or channel for 1s.