FIX: mark the cache generated by mirage calcs as CALCULATOR
The env used by mirage skills is often modified causing the resulting
env to differ. This has not caused issues to there not beign any caching
use after calcMirages but this could come up in the future. Marking the
env as CALCULATOR still allows for the env to be cached in case it's
really needed but it will prevent it being used in most cases.