Fix Volatility applying to Minion Spell skills (#8912)
* FIX(mirage): fix cap GC cooldown by mirage+attack delay

The Howlcrack unique item sets the cooldown of warcies to 0 causing the
GC dps to explode into infinity due to division by zero. This was caused
by the assumption that the mirages instantly spawn and attack. This
commit makes it so that the cooldown is at least the amount of time it
takes to spawm them all and attack or channel for 1s.

* FIX: cleanup handling of limitedSkills table

* FIX: cleanup target skill selection for SacredWisps and MirageArcher

* FIX: mark the cache generated by mirage calcs as CALCULATOR

The env used by mirage skills is often modified causing the resulting
env to differ. This has not caused issues to there not beign any caching
use after calcMirages but this could come up in the future. Marking the
env as CALCULATOR still allows for the env to be cached in case it's
really needed but it will prevent it being used in most cases.

* FIX: remove vestigial CACHE table

This was once used as the main table holding the cached output of
skills. It has been entirely replaced by the remaining tables.

* MISC: typo

* FIX: typo

* Fix Volatility applying to Minion Spell skills

Volatility was not being restricted to minion spells after we recently changed how we were applying support gems to minions

---------

Co-authored-by: Paliak <91493239+Paliak@users.noreply.github.com>
Co-authored-by: LocalIdentity <localidentity2@gmail.com>