Export 3.27 (#9142)
* Remove old dats

* New Dat columns

DefaultMonsterStats has a new column that looks like it might be a scaling column for something (Lifeblood?)
MinionType got a new bool, seems to be on vaal minions
ModType now has a new key to Stats. not sure if we care about this one
PassiveJewelRadii has a key to something new
Stats lost it's name column

* Move Clear button UI to not overlap script output

* Remove OTFile column from MonsterTypes.dat

* Fix bases export

* Export Bases

Hidden has been removed from BaseItemTypes.dat
Need to handle new `wand_can_roll_caster_modifiers` tag on wands

* Boss data

Not sure if this is correct. Would be weird to halve his damage without notice

* Cluster Export

* Crucible Export

* Essence Export

* Export Flavour Text

* Export Legion Passives

* Export Master mods

* Export minions

* Export Miscdata

* Mod Export

* Pantheon Export

* Export skill gem list

* Export Tattoos

Add parsing for the mods that changed names and the new Runegraft
Manually left in all the legacy tags on nodes as we will need another way to get them automatically now that BaseItemTypes no longer has the hidden column

* Export Stat Desscriptions

* Update existing skills

projectiles_ignore_wall_of_force
removal of `active_skill_damage` from spectres might be cause it's now the last column in DefaultMonsterStats

* Add support for new base types

Add the new wand and ring base types along with restricting their mod pool based on the implicit that block particular mods from rolling
I assumed that the mods on the rings require the corresponding element as a tag if the damage tag also exists. Will need to wait and see if this is true

* New skills

Adds support for all the new player skills
Only one not working is the Living Lightning support
Something to do with it being a effects triggered from a support is making it not show up properly

* Fix logic for new ring bases

The implicit stops any mod with a different elemental tag on it from spawning on the base

* Update jewel radius values

3.27 added new jewel radius sizes

---------

Co-authored-by: LocalIdentity <localidentity2@gmail.com>
Co-authored-by: Wires77 <Wires77@users.noreply.github.com>