Add option when using Balance of Terror to disable curse effect on enemy
When using a Balance of Terror, many users will add a curse to their build that doesn't help their damage, but give them utility in some way
The only way for the Balance of Terror mods to work are if you have the curse as an active skill in your build. Problems arise when you use 2 or more curses in your build but PoB prioritises the utility curse as the active one affecting the enemy

e.g. Use Assassins Mark + Elemental weakness in a build will always apply the Ele weakness even though you're only using it for the 40% phys as extra ele buff

I added the disable curses part to a new section atm as I wasn't too sure where to put them. Maybe the skill options part?
They only show up atm if you have the corresponding Balance of Terror mod and skill gem in a build