Add missing screen size conversions
`GetScreenSize` returns the size in physical units if the "DPI_AWARE" flag
is passed to `RenderInit`. The error popup and restart screen have the
wrong position and size at UI scaling factor != 1.0 due to missing
conversions to virtual units.

This introduces a utility function for getting the virtual screen size and
uses it to add the missing conversions.